pyproject.toml
[tool.poetry]
name = "syncer"
version = "0.1.0"
description = ""
authors = ["twilight2017 <liuzixuanp@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.9"
requests = "2.28.1"
pyyaml = "6.0"
django-ninja = "0.19.1"
loguru = "0.6.0"
djangorestframework = "3.14.0"
djangorestframework-api-key = "2.2.0"
gunicorn = "20.1.0"
gevent = "22.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
